+++ client/examples/remote_twincat.cpp
@@ -3,7 +3,7 @@
 // Created Date: 30/09/2021
 // Author: Shun Suzuki
 // -----
-// Last Modified: 30/09/2021
+// Last Modified: 21/11/2021
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2021 Hapis Lab. All rights reserved.
@@ -16,7 +16,7 @@
 
 int main() try {
   auto autd = autd::Controller::create();
-  autd->geometry()->add_device(autd::Vector3(0, 0, 0), autd::Vector3(0, 0, 0));
+  autd->geometry().add_device(autd::Vector3(0, 0, 0), autd::Vector3(0, 0, 0));
 
   const std::string remote_ipv4_addr = "your server ip";
   const std::string remote_ams_net_id = "your server ams net id";