+++ client/include/autd3/gain/backend.hpp
@@ -3,7 +3,7 @@
 // Created Date: 10/09/2021
 // Author: Shun Suzuki
 // -----
-// Last Modified: 11/09/2021
+// Last Modified: 21/11/2021
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2021 Hapis Lab. All rights reserved.
@@ -34,25 +34,25 @@ class Backend {
   Backend(Backend&& obj) = default;
   Backend& operator=(Backend&& obj) = default;
 
-  virtual void sdp(const core::GeometryPtr& geometry, const std::vector<core::Vector3>& foci, const std::vector<complex>& amps, double alpha,
+  virtual void sdp(const core::Geometry& geometry, const std::vector<core::Vector3>& foci, const std::vector<complex>& amps, double alpha,
                    double lambda, size_t repeat, bool normalize, std::vector<core::DataArray>& dst) = 0;
-  virtual void evd(const core::GeometryPtr& geometry, const std::vector<core::Vector3>& foci, const std::vector<complex>& amps, double gamma,
+  virtual void evd(const core::Geometry& geometry, const std::vector<core::Vector3>& foci, const std::vector<complex>& amps, double gamma,
                    bool normalize, std::vector<core::DataArray>& dst) = 0;
-  virtual void naive(const core::GeometryPtr& geometry, const std::vector<core::Vector3>& foci, const std::vector<complex>& amps,
+  virtual void naive(const core::Geometry& geometry, const std::vector<core::Vector3>& foci, const std::vector<complex>& amps,
                      std::vector<core::DataArray>& dst) = 0;
-  virtual void gs(const core::GeometryPtr& geometry, const std::vector<core::Vector3>& foci, const std::vector<complex>& amps, size_t repeat,
+  virtual void gs(const core::Geometry& geometry, const std::vector<core::Vector3>& foci, const std::vector<complex>& amps, size_t repeat,
                   std::vector<core::DataArray>& dst) = 0;
-  virtual void gspat(const core::GeometryPtr& geometry, const std::vector<core::Vector3>& foci, const std::vector<complex>& amps, size_t repeat,
+  virtual void gspat(const core::Geometry& geometry, const std::vector<core::Vector3>& foci, const std::vector<complex>& amps, size_t repeat,
                      std::vector<core::DataArray>& dst) = 0;
-  virtual void lm(const core::GeometryPtr& geometry, const std::vector<core::Vector3>& foci, const std::vector<complex>& amps, double eps_1,
+  virtual void lm(const core::Geometry& geometry, const std::vector<core::Vector3>& foci, const std::vector<complex>& amps, double eps_1,
                   double eps_2, double tau, size_t k_max, const std::vector<double>& initial, std::vector<core::DataArray>& dst) = 0;
-  virtual void gauss_newton(const core::GeometryPtr& geometry, const std::vector<core::Vector3>& foci, const std::vector<complex>& amps, double eps_1,
+  virtual void gauss_newton(const core::Geometry& geometry, const std::vector<core::Vector3>& foci, const std::vector<complex>& amps, double eps_1,
                             double eps_2, size_t k_max, const std::vector<double>& initial, std::vector<core::DataArray>& dst) = 0;
-  virtual void gradient_descent(const core::GeometryPtr& geometry, const std::vector<core::Vector3>& foci, const std::vector<complex>& amps,
-                                double eps, double step, size_t k_max, const std::vector<double>& initial, std::vector<core::DataArray>& dst) = 0;
-  virtual void apo(const core::GeometryPtr& geometry, const std::vector<core::Vector3>& foci, const std::vector<complex>& amps, double eps,
+  virtual void gradient_descent(const core::Geometry& geometry, const std::vector<core::Vector3>& foci, const std::vector<complex>& amps, double eps,
+                                double step, size_t k_max, const std::vector<double>& initial, std::vector<core::DataArray>& dst) = 0;
+  virtual void apo(const core::Geometry& geometry, const std::vector<core::Vector3>& foci, const std::vector<complex>& amps, double eps,
                    double lambda, size_t k_max, std::vector<core::DataArray>& dst) = 0;
-  virtual void greedy(const core::GeometryPtr& geometry, const std::vector<core::Vector3>& foci, const std::vector<complex>& amps, size_t phase_div,
+  virtual void greedy(const core::Geometry& geometry, const std::vector<core::Vector3>& foci, const std::vector<complex>& amps, size_t phase_div,
                       std::vector<core::DataArray>& dst) = 0;
 };